Extract BeaconBlockBody interface
