+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/ssz_static/SszTestExecutorDeprecated.java
@@ -29,7 +29,6 @@ import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlockHeader;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.datastructures.operations.AggregateAndProof;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.AttestationData;
@@ -70,9 +69,6 @@ public class SszTestExecutorDeprecated<T extends SszData> implements TestExecuto
               "ssz_static/AttesterSlashing",
               new SszTestExecutorDeprecated<>(AttesterSlashing.SSZ_SCHEMA))
           .put("ssz_static/BeaconBlock", new SszTestExecutorDeprecated<>(BeaconBlock.SSZ_SCHEMA))
-          .put(
-              "ssz_static/BeaconBlockBody",
-              new SszTestExecutorDeprecated<>(BeaconBlockBody.SSZ_SCHEMA))
           .put(
               "ssz_static/BeaconBlockHeader",
               new SszTestExecutorDeprecated<>(BeaconBlockHeader.SSZ_SCHEMA))
