+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/BeaconBlockBody.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 ConsenSys AG.
+ * Copyright 2021 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -13,7 +13,6 @@
 
 package tech.pegasys.teku.spec.datastructures.blocks.blockbody;
 
-import java.util.function.Function;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.bls.BLSSignature;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
@@ -22,110 +21,23 @@ import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
 import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
-import tech.pegasys.teku.ssz.SSZTypes.SSZBackingList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
-import tech.pegasys.teku.ssz.backing.SszList;
-import tech.pegasys.teku.ssz.backing.SszVector;
-import tech.pegasys.teku.ssz.backing.containers.Container8;
-import tech.pegasys.teku.ssz.backing.tree.TreeNode;
-import tech.pegasys.teku.ssz.backing.view.SszPrimitives.SszByte;
-import tech.pegasys.teku.ssz.backing.view.SszPrimitives.SszBytes32;
-import tech.pegasys.teku.ssz.backing.view.SszUtils;
-import tech.pegasys.teku.util.config.SpecDependent;
+import tech.pegasys.teku.ssz.backing.SszContainer;
 
-/** A Beacon block body */
-public class BeaconBlockBody
-    extends Container8<
-        BeaconBlockBody,
-        SszVector<SszByte>,
-        Eth1Data,
-        SszBytes32,
-        SszList<ProposerSlashing>,
-        SszList<AttesterSlashing>,
-        SszList<Attestation>,
-        SszList<Deposit>,
-        SszList<SignedVoluntaryExit>> {
+public interface BeaconBlockBody extends SszContainer {
+  BLSSignature getRandao_reveal();
 
-  // TODO(#3648) - remove this field (version schemas that depend on this field through
-  // SchemaDefinitions)
-  @Deprecated
-  public static final SpecDependent<BeaconBlockBodySchemaPhase0> SSZ_SCHEMA =
-      SpecDependent.of(BeaconBlockBodySchemaPhase0::create);
+  Eth1Data getEth1_data();
 
-  private BLSSignature randaoRevealCache;
+  Bytes32 getGraffiti();
 
-  @Deprecated
-  public BeaconBlockBody() {
-    super(SSZ_SCHEMA.get());
-  }
+  SSZList<ProposerSlashing> getProposer_slashings();
 
-  BeaconBlockBody(BeaconBlockBodySchemaPhase0 type) {
-    super(type);
-  }
+  SSZList<AttesterSlashing> getAttester_slashings();
 
-  BeaconBlockBody(BeaconBlockBodySchemaPhase0 type, TreeNode backingNode) {
-    super(type, backingNode);
-  }
+  SSZList<Attestation> getAttestations();
 
-  BeaconBlockBody(
-      BeaconBlockBodySchemaPhase0 type,
-      SszVector<SszByte> randao_reveal,
-      Eth1Data eth1_data,
-      SszBytes32 graffiti,
-      SszList<ProposerSlashing> proposer_slashings,
-      SszList<AttesterSlashing> attester_slashings,
-      SszList<Attestation> attestations,
-      SszList<Deposit> deposits,
-      SszList<SignedVoluntaryExit> voluntary_exits) {
-    super(
-        type,
-        randao_reveal,
-        eth1_data,
-        graffiti,
-        proposer_slashings,
-        attester_slashings,
-        attestations,
-        deposits,
-        voluntary_exits);
-  }
+  SSZList<Deposit> getDeposits();
 
-  public BLSSignature getRandao_reveal() {
-    if (randaoRevealCache == null) {
-      randaoRevealCache = BLSSignature.fromBytesCompressed(SszUtils.getAllBytes(getField0()));
-    }
-    return randaoRevealCache;
-  }
-
-  public Eth1Data getEth1_data() {
-    return getField1();
-  }
-
-  public Bytes32 getGraffiti() {
-    return getField2().get();
-  }
-
-  public SSZList<ProposerSlashing> getProposer_slashings() {
-    return new SSZBackingList<>(
-        ProposerSlashing.class, getField3(), Function.identity(), Function.identity());
-  }
-
-  public SSZList<AttesterSlashing> getAttester_slashings() {
-    return new SSZBackingList<>(
-        AttesterSlashing.class, getField4(), Function.identity(), Function.identity());
-  }
-
-  public SSZList<Attestation> getAttestations() {
-    return new SSZBackingList<>(
-        Attestation.class, getField5(), Function.identity(), Function.identity());
-  }
-
-  public SSZList<Deposit> getDeposits() {
-    return new SSZBackingList<>(
-        Deposit.class, getField6(), Function.identity(), Function.identity());
-  }
-
-  public SSZList<SignedVoluntaryExit> getVoluntary_exits() {
-    return new SSZBackingList<>(
-        SignedVoluntaryExit.class, getField7(), Function.identity(), Function.identity());
-  }
+  SSZList<SignedVoluntaryExit> getVoluntary_exits();
 }
