+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/BeaconBlockBodyPhase0.java
@@ -1,0 +1,140 @@
+/*
+ * Copyright 2019 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.spec.datastructures.blocks.blockbody;
+
+import java.util.function.Function;
+import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.bls.BLSSignature;
+import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
+import tech.pegasys.teku.spec.datastructures.operations.Attestation;
+import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
+import tech.pegasys.teku.spec.datastructures.operations.Deposit;
+import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
+import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
+import tech.pegasys.teku.ssz.SSZTypes.SSZBackingList;
+import tech.pegasys.teku.ssz.SSZTypes.SSZList;
+import tech.pegasys.teku.ssz.backing.SszList;
+import tech.pegasys.teku.ssz.backing.SszVector;
+import tech.pegasys.teku.ssz.backing.containers.Container8;
+import tech.pegasys.teku.ssz.backing.tree.TreeNode;
+import tech.pegasys.teku.ssz.backing.view.SszPrimitives.SszByte;
+import tech.pegasys.teku.ssz.backing.view.SszPrimitives.SszBytes32;
+import tech.pegasys.teku.ssz.backing.view.SszUtils;
+import tech.pegasys.teku.util.config.SpecDependent;
+
+/** A Beacon block body */
+public class BeaconBlockBodyPhase0
+    extends Container8<
+        BeaconBlockBodyPhase0,
+        SszVector<SszByte>,
+        Eth1Data,
+        SszBytes32,
+        SszList<ProposerSlashing>,
+        SszList<AttesterSlashing>,
+        SszList<Attestation>,
+        SszList<Deposit>,
+        SszList<SignedVoluntaryExit>>
+    implements BeaconBlockBody {
+
+  // TODO(#3648) - remove this field (version schemas that depend on this field through
+  // SchemaDefinitions)
+  @Deprecated
+  public static final SpecDependent<BeaconBlockBodySchemaPhase0> SSZ_SCHEMA =
+      SpecDependent.of(BeaconBlockBodySchemaPhase0::create);
+
+  private BLSSignature randaoRevealCache;
+
+  @Deprecated
+  public BeaconBlockBodyPhase0() {
+    super(SSZ_SCHEMA.get());
+  }
+
+  BeaconBlockBodyPhase0(BeaconBlockBodySchemaPhase0 type) {
+    super(type);
+  }
+
+  BeaconBlockBodyPhase0(BeaconBlockBodySchemaPhase0 type, TreeNode backingNode) {
+    super(type, backingNode);
+  }
+
+  BeaconBlockBodyPhase0(
+      BeaconBlockBodySchemaPhase0 type,
+      SszVector<SszByte> randao_reveal,
+      Eth1Data eth1_data,
+      SszBytes32 graffiti,
+      SszList<ProposerSlashing> proposer_slashings,
+      SszList<AttesterSlashing> attester_slashings,
+      SszList<Attestation> attestations,
+      SszList<Deposit> deposits,
+      SszList<SignedVoluntaryExit> voluntary_exits) {
+    super(
+        type,
+        randao_reveal,
+        eth1_data,
+        graffiti,
+        proposer_slashings,
+        attester_slashings,
+        attestations,
+        deposits,
+        voluntary_exits);
+  }
+
+  @Override
+  public BLSSignature getRandao_reveal() {
+    if (randaoRevealCache == null) {
+      randaoRevealCache = BLSSignature.fromBytesCompressed(SszUtils.getAllBytes(getField0()));
+    }
+    return randaoRevealCache;
+  }
+
+  @Override
+  public Eth1Data getEth1_data() {
+    return getField1();
+  }
+
+  @Override
+  public Bytes32 getGraffiti() {
+    return getField2().get();
+  }
+
+  @Override
+  public SSZList<ProposerSlashing> getProposer_slashings() {
+    return new SSZBackingList<>(
+        ProposerSlashing.class, getField3(), Function.identity(), Function.identity());
+  }
+
+  @Override
+  public SSZList<AttesterSlashing> getAttester_slashings() {
+    return new SSZBackingList<>(
+        AttesterSlashing.class, getField4(), Function.identity(), Function.identity());
+  }
+
+  @Override
+  public SSZList<Attestation> getAttestations() {
+    return new SSZBackingList<>(
+        Attestation.class, getField5(), Function.identity(), Function.identity());
+  }
+
+  @Override
+  public SSZList<Deposit> getDeposits() {
+    return new SSZBackingList<>(
+        Deposit.class, getField6(), Function.identity(), Function.identity());
+  }
+
+  @Override
+  public SSZList<SignedVoluntaryExit> getVoluntary_exits() {
+    return new SSZBackingList<>(
+        SignedVoluntaryExit.class, getField7(), Function.identity(), Function.identity());
+  }
+}
