+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/util/GenesisGenerator.java
@@ -30,7 +30,7 @@ import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodyPhase0;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
 import tech.pegasys.teku.spec.datastructures.operations.DepositData;
 import tech.pegasys.teku.spec.datastructures.state.Fork;
@@ -58,7 +58,7 @@ public class GenesisGenerator {
   public GenesisGenerator(final SchemaDefinitions schemaDefinitions) {
     state = schemaDefinitions.getBeaconStateSchema().createBuilder();
 
-    Bytes32 latestBlockRoot = new BeaconBlockBody().hashTreeRoot();
+    Bytes32 latestBlockRoot = new BeaconBlockBodyPhase0().hashTreeRoot();
     final UInt64 genesisSlot = UInt64.valueOf(Constants.GENESIS_SLOT);
     BeaconBlockHeader beaconBlockHeader =
         new BeaconBlockHeader(
