+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/datastructures/sostests/IsVariableTest.java
@@ -25,7 +25,6 @@ import tech.pegasys.teku.spec.SpecFactory;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.datastructures.forkchoice.VoteTracker;
 import tech.pegasys.teku.spec.datastructures.operations.AggregateAndProof;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
@@ -50,7 +49,7 @@ public class IsVariableTest {
 
   static Stream<Arguments> variableSizeTypes() {
     return Stream.of(
-        Arguments.of(BeaconBlockBody.SSZ_SCHEMA.get()),
+        Arguments.of(SCHEMA_DEFINITIONS.getBeaconBlockBodySchema()),
         Arguments.of(BeaconBlock.SSZ_SCHEMA.get()),
         Arguments.of(Attestation.SSZ_SCHEMA),
         Arguments.of(AttesterSlashing.SSZ_SCHEMA),
